mnirt

+++ src/calcul/plane.c
@@ -6,7 +6,7 @@
 /*   By: eflaquet <eflaquet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/11 15:36:23 by eflaquet          #+#    #+#             */
-/*   Updated: 2023/04/15 12:31:36 by eflaquet         ###   ########.fr       */
+/*   Updated: 2023/04/17 17:35:22 by eflaquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ bool	intersection_plan(t_object *plan, t_ray *ray, double *distance)
 	if (demon > EPSILON)
 	{
 		b2 = subtract_vector(plan->center, ray->origin);
-		if (dot(b2, plan->axis) / demon >= 0)
+		if ((double)(dot(b2, plan->axis) / demon) >= 0.0)
 		{
 			*distance = dot(b2, plan->axis) / demon;
 			return (true);
